load js on correct pages
