+++ core/templates/explorer/play.html
@@ -108,4 +108,7 @@
 
 {% include 'explorer/preview_pane.html' %}
 
+{% load render_bundle from webpack_loader %}
+{% render_bundle 'query' %}
+
 {% endblock content %}
