+++ core/templates/explorer/query.html
@@ -168,4 +168,7 @@
 
 {% include 'explorer/preview_pane.html' %}
 
+{% load render_bundle from webpack_loader %}
+{% render_bundle 'query' %}
+
 {% endblock content %}
